Materials: Break UnityCustomData out into separate file
