+++ unity-assetpackage/Assets/UsdUnitySdk/Scripts/IO/Materials/UnityMaterialData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8cd6aeaaa81b3b84f9b767ed3352b7e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
